add check to soundness script for even numbered versions in the package.json
